First test of dynamic table font change.

+++ gui/com/frostwire/gui/theme/ThemeMediator.java
@@ -170,7 +170,7 @@ public final class ThemeMediator {
         for (Component comp : comps) {
             if (comp instanceof LimeJTable) {
                 Font f = comp.getFont();
-                f = f.deriveFont(f.getSize() + delta);
+                f = f.deriveFont((float)(f.getSize() + delta));
                 UIDefaults nimbusOverrides = new UIDefaults();
                 nimbusOverrides.put("Table.font", f);
                 ((LimeJTable) comp).putClientProperty("Nimbus.Overrides", nimbusOverrides);
